is: Move SetContext funcs to single file
