+++ pkg/identityserver/store/user.go
@@ -15,7 +15,6 @@
 package store
 
 import (
-	"context"
 	"time"
 
 	"github.com/gogo/protobuf/types"
@@ -59,12 +58,6 @@ type User struct {
 
 func init() {
 	registerModel(&User{})
-}
-
-// SetContext sets the context on both the Model and Account.
-func (usr *User) SetContext(ctx context.Context) {
-	usr.Model.SetContext(ctx)
-	usr.Account.SetContext(ctx)
 }
 
 // functions to set fields from the user model into the user proto.
